refactor: rearrange folder structure
