--- src/common/components/index.ts
@@ -1,6 +1,0 @@
-export * from './Logo';
-export * from './NavLink';
-export * from './Navbar';
-export * from './Layout';
-export * from './JobsFilter';
-export * from './SearchBar';
