+++ src/components/search/JobsFilter/IndustryFilter/IndustryFilter.module.scss
