+++ src/components/search/JobsFilter/SalaryRangeFilter/SalaryRangeFilter.module.scss
