+++ src/components/search/JobsFilter/SalaryRangeFilter/SalaryRangeFilter.tsx
