+++ src/components/search/JobsFilter/SalaryRangeFilter/index.ts
