+++ src/components/search/SearchBar/MagnifierIcon/MagnifierIcon.module.scss
