+++ src/components/search/SearchBar/MagnifierIcon/MagnifierIcon.tsx
