+++ src/components/search/SearchBar/MagnifierIcon/index.ts
