+++ src/components/search/index.ts
@@ -1,0 +1,2 @@
+export * from './JobsFilter';
+export * from './SearchBar';
