+++ src/stories/Navbar.stories.ts
@@ -1,4 +1,4 @@
-import { Navbar } from 'common/components';
+import { Navbar } from 'components/common/Layout/Navbar';
 
 export default {
   title: 'Components/Navbar',
